requirements/base.txt
Django==2.1.7
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
requests==2.20.1
Pillow==5.4.1
boto3==1.9.112
psycopg2-binary==2.7.7
django-cors-headers==2.5.2
pika==1.0.0
django-notifs==2.6.2
httplib2==0.12.1
google-api-python-client==1.7.8
oauth2client==4.1.3
spacy==2.1.3
celery==4.3.0
pandas==0.25.3
jellyfish==0.7.1
dhooks==1.1.0
discord-webhook==0.4.1
python-dotenv==0.10.3
xlrd==1.2.0
django-sql-explorer==1.1.3
gunicorn==19.9.0
gevent==1.4.0
django-storages==1.8
pyfcm==1.4.7
django-constance[database]
environs==7.2.0
